stage2: add call_async ZIR instruction
This is to be used when the source code looks like this:

    var a = async b();

The instruction acts both as an alloc as well as a function call using
that alloc as the result location. This avoids multiple ZIR instructions
as well as complicated "inferred pointer" semantic analysis for a common
case.

AstGen is not yet updated to emit this new instruction.